refactor: Merge each tool's *_common.ts into its tool-entry file
Each of the four domain tools had its logic split across a `*_common.ts`
sibling. The split made sense when the files lived in separate directories
(default/ vs core/); now that both sit in the same domain dir the suffix
is noise.

Fold each `*_common.ts` into its tool-entry file and delete the source
file. Update all import sites (widgets, server.ts, actor_executor.ts,
tests). No behavior change — exports keep identical names and signatures.

The only non-trivial adjustment: `fetchActorDetails` imported from
`utils/actor_details.js` was aliased to `fetchActorDetailsFromApi` to
avoid a name collision with the `fetchActorDetails: ToolEntry` export
now defined in the same file.

Closes #1020
